Output appropriate client library namespace. (#2)
This commit makes it such that library code is output into an appropriate import structure.

Prior to this, each service was written to the output directory at the top level. Now, the API's namespace and name are used. Therefore, for an API named "Foo" with version "v1", which has `service Bar`...

  * Before: `import bar`
  * Now: `from foo_v1 import bar`

And if the namespace is `("Spam", "Eggs")`...

  * Before: `import bar`
  * Now: `from spam.eggs.foo_v1 import bar`

Additionally, the API's name is directly importable, with and without the version:

  * Now: `from spam.eggs import foo`
  * Now: `from spam.eggs import foo_v1`

These are actually the preferred import approach.

Internally, this commit changes the file rename approach from just renaming filenames with templates beginning with `service/` to now substituting four variables (as documented).