Refactor the wrap filter.
This replaces the `initial_width` and `subsequent_indent` keyword arguments
with a single `offset` argument. The `offset` can be a int or a two-tuple of
ints (first line and subsequent lines, respectively), and the behavior is the
same.

This does remove the ability to have a `subsequent_indent` which is anything
other than spaces, which did not end up being used in real life (at least not yet).
If it is needed, the `subsequent_indent` stand-alone filter could be resurrected
(see #14).

This change markedly improves readability of templates where `wrap` is being used.
Similarly, the syntactic change from "initial width" to "initial offset" ends
up being far more intuitive (I had used "initial width" incorrectly about half
the time and had to debug it).