Add simple whitespace post-processing.
This commit adds very basic whitespace formatting to nullify
certain whitespace issues that are sufficiently difficult to get
right in Jinja (particularly when `if` and `for` are involved).

This is not meant as a replacement for `yapf`. It is completely
reasonable to want to run the resulting code through `yapf`,
but I do not want to force this on all template authors. This
commit is more humble and just applies very specific fixes
that are more or less universally adopted and sufficiently
difficult to do.