Table-setting for generating protos.
This commit makes several schema edits and additions in order to facilitate
generating pb2 replacements directly.

In particular, it:

  * Adds support for a `$proto` rewrite for filenames (such files
    receive a `proto` variable), and adds loader support for such.
  * Refactors `Address` to include the `name` of the object it is
    attached to, enabling it to be used as the `ident` object.
  * Refactors the various `*_ident` properties into the address,
    and re-orders template use (e.g. `sphinx_ident` is now
    `ident.sphinx`).
      * Additionally, `python_ident` is now just `ident` in templates.
  * Adds a `rel` method for getting identifiers relative to the
    file being generated. This deals with a problem that will surface
    in replacing pb2s where a file would attempt to import or reference
    itself.

LROs are slightly broken in a slightly different way than before
(note that they were broken in the status quo ante due to a bug
in `OperationType`), and `OperationType` turned out to be a mistake.
Filed #34 to track this, and it will be fixed after pb2 replacements
are introduced (since they will significantly alter what the fix is).