[refactor] Change library structure, import everything into root module.
This commit changes the structure of the resulting libraries to be
more consistent when subpackages are in use: everything (both services
and messages) goes under the subpackage:

```
$name_$version/
  $sub/
    services/
    types/
```

Imports are changed so that subpackages are imported directly into the
root module, and all services and messages are imported directly into
either the root module (if they are not in a subpackage) or the
subpackage module (if they are).

The new structure is simpler (everything works the same way) and safer
(no possibility of naming collisions on services).