[docs] Fix access denied error for image resource in verifying library section
The example script for verifying the library over
here: https://gapic-generator-python.readthedocs.io/en/stable/getting-started/local.html#verifying-the-library
has a link to a resource whose acces is denied which results in the
following error on running :  "The URL does not appear to be accessible
by us. Please double check or download the content and pass it in."
The new link added gives a proper response from the service on running.
Output can be found here : https://pastebin.com/U0XGFcrt