[fix] Do not silently drop `--python-gapic-` options.
This is a follow-up to #126, where I added a CLI
option, but neglected to update the Docker
entrypoint script to actually pass it on to the
generator, making the option useless when using
Docker (the common case for now).

This updates the script to be more robust, and
check the prefix rather than an index of every
known option.