[feat] Add mypy checks to the generated output.
Many things can not be typechecked yet due to Python 2 compatibility.
We will come back to those as we move everything to Python 3.5+.

We also do not yet have exported type-checking in proto-plus; I
will make a subsequent PR to remove the `#type: ignore` from
that import once proto-plus type checking is fully in place and
exported.