All response variable/attribute expressions are validated
All statements and expressions that reference the method response are validated,
e.g. {"print": ["%s", "$resp.cephalopods[0].mantle"]},
will be checked to make sure that the response message type has a
repeated "cephalopods" field, and that that type has a "mantle" field.

Includes a major refactor of Validator.validate_expression