Replace every $ in all template file paths with % (#247)
Unix utilities and editors attempt to expand any name beginning with $
as if it were a variable; if the name does not reference a previously
defined variable, it is given a default (empty) value.

This is profoundly irritating when trying to edit template files in a
command-line environment. As a fix, the naming convention is changed
to use % for template filename/filepath intperpolation.