Add generated unit test coverage for lazy import, part 1
The Python 3.7 showcase unit tests (and all 'real' generated unit
tests when using the lazy import feature) include a test to verify
that the logic used for lazy import is covered.

If the classes and submodules are available it is assumed that they
are the correct ones; asserting that there is no mismatch between lazy
import path and corresponding object is outside the current scope of
these tests.