Cross package request messages are constructed specially
A cross package request is almost certainly not a proto-plus wrapped
type, which puts certain restrictions on its construction and
manipulation due to constraints in the Python protobuf API.

It is bad practice but neither forbidden nor impossible to write a
method whose request message definitionlives in a different
package. A recurring example is IAM Policy Requests.

This change detects when a method's request lives in a different
package and constructs it either via keyword expansion (a dict was
passed in) or with no ctor params.

Also, in the above scenario, fields are not eligible for flattening if
they point to non-primitive types.