Fix for flattened unit tests with deep nesting
Expands the unit-test visible method ref types to include _all_
recursive types.

Do NOT use this deep nesting in the client submodule. The client ONLY
needs to import, for each method, the request type, response type, and
the types of any flattened fields.

I swear, most of this is really necessary.