Move `libcst` to `install_requires` (#423)
Use of `setup_requires` is discouraged by [setuptools](https://setuptools.readthedocs.io/en/latest/setuptools.html?highlight=setup_requires#options)


A string or list of strings specifying what other distributions need to be present in order for the setup script to run. setuptools will attempt to obtain these (even going so far as to download them using EasyInstall) before processing the rest of the setup script or commands. This argument is needed if you are using distutils extensions as part of your build process; for example, extensions that process setup() arguments and turn them into EGG-INFO metadata files.

(Note: projects listed in setup_requires will NOT be automatically installed on the system where the setup script is being run. They are simply downloaded to the ./.eggs directory if they’re not locally available already. If you want them to be installed, as well as being available when the setup script is run, you should add them to install_requires and setup_requires.)


`libcst` is also technically not needed by the setup script, so adding it as an `install_requires` feels more appropriate.