fix: resource messages in method response types generate helpers
Some resource messages are only referenced in method responses, either
directly (the method returns a resource) or indirectly (the resource
is a field for some other message).
These response-resources now generate helper methods in client
classes.

Contains a minor formatting fix in the generated output, a minor fix
in an error message, and a tweak to nox.py to aid interactive debugging.