fix: disable always_use_jwt_access
Some APIs with Cloud Storage integrations seem to have failing
samples when non-default scopes are provided. This PR disables the feature
globally for now while we investigate.

We can no longer rollback da119c72c8 as some libraries have been
released with the change and removing the kwarg would be a breaking
change.