feat: add interceptor-like functionality to REST transport
Interceptors are a gRPC feature that wraps rpcs in
continuation-passing-style pre and post method custom functions.
These can be used e.g. for logging, local caching, and tweaking
metadata.

This PR adds interceptor like functionality to the REST transport in
generated GAPICs.

The REST transport interceptors differ in a few ways:
1) They are not continuations. For each method there is a slot for a
"pre"function, and for each method with a non-empty return there is a
slot for a "post" function.
2) There is always an interceptor for each method. The default simply
does nothing.
3) Existing gRPC interceptors and the new REST interceptors are not
composable or interoperable.