feat: full LRO for Extended Operations
Implementation and tests for Extended Operations.

The full description of Extended Operations is beyond the scope of
this changelog, but three important differences from 'normal' LROs are
1) The type of the 'Operation' is not fixed but is determined by
specification.
2) The operation service is not fixed and is determined by specification.
3) The polling method in the operation service is not fixed and is
determined by specification.

Methods that meet the Extended Operations criteria will generate two
methods in the client class: a 'plain' version with the "_unary"
suffix that returns the custom Operation type as a raw proto message,
and a 'rich' version that returns google.api_core.ExtendedOperation,
which can be polled and waited on for completion.