feat: Note that "rest" transport support is beta.
This beta notice is triggered by having numeric enum support disabled (the default), since numeric enum support is currently the only blocker to supporting REST transport fully.

This condition can be changed if further blockers arise to a GA release of REST transport.

This logic and notice can be removed entirely once we are satisfied that REST transport works with or without numeric enums.